+++ randomgen/speck128.pyx
@@ -47,8 +47,8 @@ cdef class SPECK128(BitGenerator):
     mode : {None, "sequence", "legacy"}, optional
         The seeding mode to use. "legacy" uses the legacy
         SplitMix64-based initialization. "sequence" uses a SeedSequence
-        to transforms the seed into an initial state. None defaults to "legacy"
-        and warns that the default after 1.19 will change to "sequence".
+        to transforms the seed into an initial state.  None defaults to
+        "sequence".
 
     Attributes
     ----------